% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salary_model.R
\docType{data}
\name{salary_model}
\alias{salary_model}
\title{Salary Estimating model}
\format{
a \code{\link{gam}} object
}
\source{
Udemy - Superdatascience
}
\usage{
salary_model
}
\description{
Simple example model that predicts salary from Age and Gender
}
